2/25 update
Played around with this a little more.  Shortened the code by throwing away the increment function.  Really wasn't needed here - and it seems to run flawlessly.  May come back and try to figure out a better way to discern the else if / or combo at a later time.  This should be my final commit.